project3/requirements.txt
streamlit
langchain
huggingface_hub
python-dotenv
tiktoken
faiss-cpu